vnstat_assist/python/requirements.txt
# requirements.txt 需要添加：
requests>=2.26.0
python-dateutil>=2.8.2
apscheduler==3.10.4